services/voice-service/pyproject.toml
# =============================================================================
# SCHOLARLY VOICE SERVICE
# =============================================================================
# Self-hosted TTS, STT, Voice Cloning & Educator Studio Tools
# A standalone Python microservice that provides voice AI capabilities
# through a RESTful API, designed to operate independently while
# integrating seamlessly into the Scholarly ecosystem via AIPAL.
# =============================================================================

[project]
name = "scholarly-voice-service"
version = "0.1.0"
description = "Self-hosted voice AI service for the Scholarly educational platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Scholarly", email = "dev@scholarly.app" }
]

dependencies = [
    # --- HTTP Framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # --- TTS Models ---
    "kokoro>=1.0.0",
    # chatterbox-tts installed separately (Sprint 29 Week 4)

    # --- STT Models (Sprint 29 Week 2) ---
    # "faster-whisper>=1.1.0",
    # "whisperx>=3.0",

    # --- Audio Processing ---
    "librosa>=0.11.0",
    "pyrubberband>=0.4.0",
    "pyloudnorm>=0.1.0",
    "noisereduce>=3.0",
    "soundfile>=0.12.0",
    "scipy>=1.0",
    "numpy>=1.0",

    # --- ML Runtime ---
    "torch>=2.0",

    # --- Storage ---
    "azure-storage-blob>=12.0",

    # --- Caching ---
    "redis>=5.0",

    # --- Observability ---
    "prometheus-client>=0.20.0",

    # --- HTTP Client (for webhook emission & health checks) ---
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

# STT dependencies separated so TTS-only deployments stay lean
stt = [
    "faster-whisper>=1.1.0",
    # whisperx requires manual install due to torch version constraints
]

# Voice cloning dependencies
cloning = [
    "chatterbox-tts",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["app*", "providers*", "processing*", "cloning*", "storage*", "models*"]

# --- Tooling Configuration ---

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require models loaded)",
    "gpu: Tests requiring GPU (skip on CPU-only runners)",
    "slow: Tests that take >10s (model loading, full pipeline)",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
